requirements.txt
decorator==4.4.2
et-xmlfile==1.1.0
future==0.18.2
networkx==2.5.1
numpy==1.20.3
openpyxl==3.0.9
pandas==1.2.4
progressbar33==2.4
psutil==5.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
treelib==1.6.1